requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
future==1.0.0
fvcore==0.1.5.post20221221
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.3.0
portalocker==3.2.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pyFFTW==0.14.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
sigpy==0.1.27
six==1.17.0
sporco==0.2.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
tifffile==2024.8.30
timm==1.0.16
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
urllib3==2.5.0
yacs==0.1.8
zipp==3.23.0
